+++ CHANGES.md
@@ -4,6 +4,8 @@ The release versions are PyPi releases.
 ## Version 3.3 (as yet unreleased)
 
 #### New Features
+  * Added possibility to patch modules that import file system modules under another name, 
+  e.g. `import os as '_os` ([#231](../../issues/231)).
   * Added support for `dir_fd` argument in several `os` functions ([#206](../../issues/206)). 
   * Added support for open file descriptor as path argument
     in `os.utime`, `os.chmod`, `os.chdir`, `os.chown`, `os.listdir`, `os.stat` and `os.lstat`